[cleanup] removed JSONAssert due to licensing issues
We cannot use JSONAssert anymore, because of licensing issues. For
this to work we had to ignore many of the important JSON tests. These
tests have to be re-activated when:

  1) JSONAssert fixes its licensing issues (see https://github.com/skyscreamer/JSONassert/issues/44)
  2) An alternative library for comparing JSON has been found

Beware of the following: Whenever you change things in this project
that might affect the generated JSON you HAVE TO re-activate the
test, either using method 1), 2), or implementing an own JSON
test mechanism!