Deprecate User/Group/Resource#setSchemas(...)
Setting the schemata is totally senseless. If no extensions are involved,
then adding another schema or even overwrite the core schemata, would
mean you have a different resource, i.e. not a `User` but an
`EnterpriseUser`, which is not possible, without adding new fields, etc.
to the existing classes. If extensions are involved, then adding the
extension will automatically add the URN to the list of schemata.

This commit deprecates these methods, including a hint, that
they will be finally removed in version 1.8 or 2.0. Three versions
should be enough for users to update. Hopefully, nobody is using the
method anyway.

Resolves #127